+++ examples/example_vanilla_FDN.py
@@ -59,8 +59,8 @@ def _(pyFDN, torch):
         direct_gain=1.0,
         rt=2.0,
         rt_nyquist=0.5,
-        eq_db_dc=0.0,
-        eq_db_nyquist=-6.0,
+        output_gain_db=0.0,
+        output_gain_db_nyquist=-6.0,
         rng=42,
     )
     model = pyFDN.dss_to_flamo(